requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas 
numpy
python-dotenv
sentence_transformers
dotenv
streamlit